Ecommerce-Analytics/pyproject.toml
[project]
name = "end-to-end-e-commerce-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "apache-airflow==2.8.1",
    "dbt-core==1.9.0",
    "dbt-snowflake>=1.9.4",
    "faker>=40.12.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "snowflake-connector-python[pandas]>=3.18.0",
    "snowflake-sqlalchemy>=1.9.0",
    "sqlalchemy>=1.4.54",
]
